graphql_server/pyproject.toml
[tool.poetry]
name = "graphql-server"
version = "0.1.0"
description = ""
authors = ["Nazar Vinnichuk <nazar@vinnich.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["debug-server"], version = "^0.209.5"}
swagger-client = {path = "../openapi_client/", develop = true}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
